Adds README file
Adds a README file to provide a basic description of the extension.

It explains the extension's purpose, which is to change the user's theme based on the accent color chosen in Gnome Settings.
It also mentions the extension's origin as a modification of another project.
Finally, it outlines a future task to enable theme changes based on the light/dark color scheme.